+++ src/Extension/FrontMatter/FrontMatterParser.php
@@ -50,8 +50,12 @@ final class FrontMatterParser implements FrontMatterParserInterface
         $data = $this->frontMatterParser->parse($frontMatter);
 
         // Advance through any remaining newlines which separated the front matter from the Markdown text
-        $cursor->match('/^\n+/');
+        $trailingNewlines = $cursor->match('/^\n+/');
 
-        return new MarkdownInputWithFrontMatter($cursor->getRemainder(), $data);
+        // Calculate how many lines the Markdown is offset from the front matter by counting the number of newlines
+        // Don't forget to add 1 because we stripped one out when trimming the trailing delims
+        $lineOffset = \preg_match_all('/\n/', $frontMatter . $trailingNewlines) + 1;
+
+        return new MarkdownInputWithFrontMatter($cursor->getRemainder(), $lineOffset, $data);
     }
 }