+++ src/Extension/FrontMatter/Input/MarkdownInputWithFrontMatter.php
@@ -23,11 +23,12 @@ final class MarkdownInputWithFrontMatter extends MarkdownInput implements FrontM
 
     /**
      * @param string     $content     Markdown content without the raw front matter
+     * @param int        $lineOffset  Line offset (based on number of front matter lines removed)
      * @param mixed|null $frontMatter Parsed front matter
      */
-    public function __construct(string $content, $frontMatter = null)
+    public function __construct(string $content, int $lineOffset = 0, $frontMatter = null)
     {
-        parent::__construct($content);
+        parent::__construct($content, $lineOffset);
 
         $this->frontMatter = $frontMatter;
     }