+++ CHANGELOG.md
@@ -102,6 +102,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
    - Footnotes can now contain multiple lines of content, including sub-blocks, by indenting them
    - Footnote event listeners now have numbered priorities (but still execute in the same order)
    - Footnotes must now be separated from previous content by a blank line
+ - The line numbers (keys) returned via `MarkdownInput::getLines()` now start at 1 instead of 0
 
 ### Fixed